allows specification of levels when predicting
Since I have not found a simple way of accurately computing the number of
levels for a distributed solve serially, this allows the user to specify
the number of levels when predicting the best number of blocks for
redistribution.  By allowing this we will at least be accurate for the
finest level of redistribution.