Temporary bugfix for factorization hack in tridiagonal solver.
The multilevel tridiagonal solver includes a hack to avoid repeating
the factorization in Phase A.  This hack uses the Fortran SAVE
statement with a fixed array which limits the number of levels that
can be used.  This commit fixes the hack so the line relaxation works
when performed in a red-black ordering.