Simplify type deduction for older compilers, improve travis.
This includes a small change to perf_factory to pass a stl function
explicitly instead of a lambda which simplifies type deduction.  This
was needed for earlier C++11 compilers.  This commit also includes an
update to the build system to automatically download/build gtest if it
is not found on the system.