Update 3D code after 2D refactoring.
The 2D code was refactored to abstract the halo exchange layer.  This
was done so that the halo exchange could be chosen by the user and to
ease Tausch integration.  The refactoring also included using the
redistribution library to support a generic inner solve so that it
could be used for the BoxMG and LU coarse solve.  Using the
redistribution library for these coarse solves removed code
duplication in the Fortran kernels.  This commit updates the 3D code
with the above improvements.